extracted out registration details and patient details to separate file
