+++ nhais-adaptor/adaptor/outgoing/tests/test_message_adaptor.py
@@ -8,8 +8,9 @@ from fhirclient.models.humanname import HumanName
 from fhirclient.models.address import Address
 from edifact.outgoing.models.name import Name
 from edifact.outgoing.models.address import Address as EdifactAddress
-from edifact.outgoing.models.message import MessageSegmentBirthPatientDetails, MessageSegmentBirthRegistrationDetails, \
-    MessageBeginning, Message
+from edifact.outgoing.models.message import MessageBeginning, Message
+from edifact.outgoing.models.message_segment_patient_details import MessageSegmentBirthPatientDetails
+from edifact.outgoing.models.message_segment_registration_details import MessageSegmentBirthRegistrationDetails
 
 
 class TestMessageAdaptor(unittest.TestCase):
